# fix compat mode perf_genericevents