added code to run script correctly in compat mode instead of
power model using pvr values

Signed-off-by: Disha Goel <disgoel@linux.vnet.ibm.com>

before patch getting this error in compat mode:
10:48:23 INFO    : Running: /usr/local/bin/avocado run --nrunner-max-parallel-tasks=1 /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_genericevents.py --force-job-id 6519e05c34f50ae45126fd61a0e09e6c197aa512  --job-results-dir /home/avocado-fvt-wrapper/results
JOB ID     : 6519e05c34f50ae45126fd61a0e09e6c197aa512
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2022-06-29T10.48-6519e05/job.log
 (1/1) /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_genericevents.py:test_generic_events.test: STARTED
 (1/1) /home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf/perf_genericevents.py:test_generic_events.test:  WARN: Test passed but there were warnings during execution. Check the log for details. (0.06 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 1 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2022-06-29T10.48-6519e05/results.html
JOB TIME   : 2.68 s

after patch:
JOB ID     : f4eeea4b05dc4c90b1014c48f05905c11fcacc62
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2022-07-06T01.53-f4eeea4/job.log
 (1/1) perf_genericevents.py:test_generic_events.test: STARTED
 (1/1) perf_genericevents.py:test_generic_events.test: PASS (0.02 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2022-07-06T01.53-f4eeea4/results.html
JOB TIME   : 1.81 s

script works fine in normal P8/P9/P10 mode also

[debug.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/9052335/debug.log)